refactor(globals): rename third-party to github-hooks
