+++ globals/github-hooks/src/get-github-auth-token/get-github-auth-token.hook.ts
