+++ globals/github-hooks/src/get-github-auth-token/index.ts
