+++ globals/github-hooks/src/get-members-without-2fa/get-members-withot-2fa.interfaces.ts
