+++ globals/github-hooks/src/index.ts
